// ui/src/orval/models/tablePreviewDataColumn.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UI Router API
 * Defines the specification for the UI Catalog API
 * OpenAPI spec version: 1.0.2
 */
import type { TablePreviewDataRow } from './tablePreviewDataRow';

export interface TablePreviewDataColumn {
  name: string;
  rows: TablePreviewDataRow[];
}