// ui/src/orval/models/tablePreviewDataResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UI Router API
 * Defines the specification for the UI Catalog API
 * OpenAPI spec version: 1.0.2
 */
import type { TablePreviewDataColumn } from './tablePreviewDataColumn';

export interface TablePreviewDataResponse {
  items: TablePreviewDataColumn[];
}